requirements.txt
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.12
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
tqdm>=4.66.0
pillow>=10.0.0
flask>=3.0.0
werkzeug>=3.0.0
